% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execSQL.R
\name{execSQL}
\alias{execSQL}
\title{Execute query in database}
\usage{
execSQL(sql, replacement = NULL, connection_alias = "default")
}
\arguments{
\item{sql}{Query SQL to be executed: 'INSERT INTO cars (speed, dist) VALUES (1, 1), (2, 2), (3, 3);'}

\item{replacement}{Array with replecements c('find this' = 'replace with this')}

\item{connection_alias}{Alias para acesso a conexao armazenada}
}
\value{
TRUE if executed without errors
}
\description{
Execute query in database
}
\author{
Herson Melo
}
