% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execStoredScript.R
\name{execStoredScript}
\alias{execStoredScript}
\title{Get Script Query stored}
\usage{
execStoredScript(script, replacement = NULL, connection_alias = "default")
}
\arguments{
\item{script}{Script name}

\item{replacement}{Array with replecements c('find this' = 'replace with this')}

\item{connection_alias}{Alias para acesso a conexao armazenada}
}
\value{
Query Executed
}
\description{
Get Script Query stored
}
\author{
Herson Melo
}
