% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_conexao.R
\name{get_conexao}
\alias{get_conexao}
\title{Consulta o respositorio local e retorna a string de conexao armazenada}
\usage{
get_conexao(name = "default")
}
\arguments{
\item{name}{Nome da conexao armazenada}
}
\value{
Dados de conexao para acesso ao banco de dados
}
\description{
Consulta o respositorio local e retorna a string de conexao armazenada
}
\author{
Herson Melo <hersonpc@gmail.com>
}
