% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_constantes.R
\name{init_constantes}
\alias{init_constantes}
\title{Inicializa um setup de constantes}
\usage{
init_constantes()
}
\description{
Inicializa um setup de constantes
}
\details{
Ativa um pacote de constantes no ambiente global
}
\examples{
init_constantes()
}
\author{
Herson Melo
}
