% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listar_conexoes.R
\name{listar_conexoes}
\alias{listar_conexoes}
\title{Lista o alias das conexoes armazenadas localmente}
\usage{
listar_conexoes()
}
\value{
Retorna relação dos nomes das conexoes armazenadas localmente
}
\description{
Lista o alias das conexoes armazenadas localmente
}
\examples{
listar_conexoes()
}
\author{
Herson Melo
}
