% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_exec.R
\name{mv_exec}
\alias{mv_exec}
\title{Executar Query no banco de dados MV}
\usage{
mv_exec(query)
}
\arguments{
\item{query}{Query SQL a ser executada}
}
\value{
Conexao ao banco
}
\description{
Exibe a mensagem de ola mundo
}
\author{
Herson Melo <hersonpc@gmail.com>
}
