% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obter_conexao.R
\name{obter_conexao}
\alias{obter_conexao}
\title{Obtem os dados da conexao armazenada localmente}
\usage{
obter_conexao(name = "default")
}
\arguments{
\item{name}{Nome (Alias) da conexao armazenada}
}
\value{
Dados de conexao para acesso ao banco de dados
}
\description{
Obtem os dados da conexao armazenada localmente
}
\examples{
obter_conexao(name = "default")
}
\author{
Herson Melo <hersonpc@gmail.com>
}
