% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\title{Executar Query no banco de dados MV}
\usage{
query(sql, replacement = NULL, connection_alias = "default")
}
\arguments{
\item{sql}{Query SQL a ser executada ex: 'select * from tabela'}

\item{replacement}{Array with replecements c('find this' = 'replace with this')}

\item{connection_alias}{Alias para acesso a conexao armazenada}
}
\value{
Query executada no banco de dados
}
\description{
Executar Query no banco de dados MV
}
\author{
Herson Melo
}
