% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_acentos.R
\name{remove_acentos}
\alias{remove_acentos}
\title{Funcao para remover os acentos}
\source{
https://pt.stackoverflow.com/questions/46473/remover-acentos
}
\usage{
remove_acentos(str, pattern = "all")
}
\arguments{
\item{str}{vetor de strings que terão seus acentos retirados.}

\item{pattern}{vetor de strings com um ou mais elementos indicando quais acentos deverão ser retirados.
Para indicar quais acentos deverão ser retirados, um vetor com os símbolos deverão ser passados.
Exemplo: pattern = c("´", "^") retirará os acentos agudos e circunflexos apenas.
Outras palavras aceitas: "all" (retira todos os acentos, que são "´", "`", "^", "~", "¨", "ç")}
}
\value{
O resultado é uma string com o texto sem acentos
}
\description{
Funcao para remover os acentos
}
\details{
Função para forçar a remocao de acentos de uma string
}
\examples{
texto <- remove_acentos("Participação")
}
