% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_conexao.R
\name{save_conexao}
\alias{save_conexao}
\title{Consulta o respositorio local e retorna a string de conexao armazenada}
\usage{
save_conexao(name = "default", stringConexao, username, password)
}
\arguments{
\item{name}{Nome da conexao salva (ex: 'default')}
}
\value{
String de conexao para acesso ao banco
}
\description{
Consulta o respositorio local e retorna a string de conexao armazenada
}
\author{
Herson Melo <hersonpc@gmail.com>
}
