% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_ambiente_oracle.R
\name{setup_ambiente_oracle}
\alias{setup_ambiente_oracle}
\title{Inicializa um setup para o ambiente Oracle}
\usage{
setup_ambiente_oracle()
}
\description{
Inicializa um setup para o ambiente Oracle
}
\details{
Inicializa um setup para o ambiente Oracle
}
\examples{
setup_ambiente_oracle()
}
\author{
Herson Melo
}
