% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_replace.R
\name{sqlReplace}
\alias{sqlReplace}
\title{Get Script Query stored}
\usage{
sqlReplace(text, replacement = NULL)
}
\arguments{
\item{text}{Original text}

\item{replacement}{Array with replecements c('find this' = 'replace with this')}
}
\value{
Replaced text
}
\description{
Get Script Query stored
}
\author{
Herson Melo
}
