% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeTable.R
\name{writeTable}
\alias{writeTable}
\title{Write data frames in database as a table}
\usage{
writeTable(
  df,
  table_name,
  delete_table_if_exists = FALSE,
  append = FALSE,
  overwrite = FALSE,
  row.names = FALSE,
  conexao_name = "default"
)
}
\arguments{
\item{df}{Tabela de dados}

\item{table_name}{Table name in database}

\item{delete_table_if_exists}{If TRUE and table exists in database, it will be deleted}

\item{append}{Append parameter for dbWriteTable}

\item{overwrite}{Overwrite parameter for dbWriteTable}

\item{row.names}{Row.names parameter for dbWriteTable}

\item{conexao_name}{Alias name for connection stored in local machine}
}
\value{
TRUE or FALSE if table has created
}
\description{
Write data frames in database as a table
}
\author{
Herson Melo
}
