% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_arquivo.R
\name{ler_arquivo}
\alias{ler_arquivo}
\title{Le um arquivo e obtem seu conteudo}
\usage{
ler_arquivo(filename)
}
\arguments{
\item{filename}{Caminho do arquivo a ser lido}
}
\value{
Conteudo do arquivo
}
\description{
Le um arquivo e obtem seu conteudo
}
\author{
Herson Melo <hersonpc@gmail.com>
}
