% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\title{Executar Query no banco de dados MV}
\usage{
query(sql, conexao_name = "default")
}
\arguments{
\item{sql}{Query SQL a ser executada ex: 'select * from tabela'}

\item{conexao_name}{Alias para acesso a conexao armazenada}
}
\value{
Query executada no banco de dados
}
\description{
Exibe a mensagem de ola mundo
}
\author{
Herson Melo <hersonpc@gmail.com>
}
